test: add comprehensive tests for API calls, user creation, and forecast insertion
